Refactor API base functions
